Added post processing rendering stage
Added functionality to the rendering stage of LightningOpenGL where the first pass renders to a renderbuffer object, the texture bound to that rbo is then rendered to a quad that covers the entire screenspace. This process provides an inbetween stage where we can alter the results before outputing them, post process effects.