Added Point Shadows, Updated Rendering Pipeline and Shaders
Added the base functionality for point lights that use omnidirectional shadow mapping, a cube map is generated for the light which is passed to the object shader. Altered the scene to demonstrate the shadows. Updated shaders to be complatible with the new rendering pipleine structure and removed shaders no longer used from LightningOpenGL.cpp. Created shader loading function in FunctionLibrary that accounts for possible geometry shaders. Added keys, B and V, to toggle blur(glow) and shadows respectively. Updated RenderCube functions verticies to have tex coords and changed the starting position of the camera