Added Dynamic and Static array types and Compatability with the rest of the program. Removed triangle bolt segments, Added GUI elements, Updated Scene and Added debug tool to LineBoltSegmenet
Added Dynamic and Static arrays for each array used to make the lightning bolt: Pattern array, line segments and light positions. Created the BoltSetup file to handle the processes of getting the lightning pattery array and using it to give values to the other arrays. LightningPatters was updated to take either dynamic or static arrays. The main rendering process was also updated to handle the changes

Removed the triangle bolt segments options, the class is still there but the Line primatives do a fine job and I don't have any plans that envolve different primatives so they are useless at the moment

Updated the method choice GUI option as well as how it interacts with the new BoltSetup file, cleaning up the process and moving it out of the main LightningOpenGL file. Added a button to toggle between Dynamic and Static arrays.

Added printInfo function to LineBoltSegment, it's a sometimes useful debugging tool. Removed the color variable from LineBoltSegment as it served no purpose