Added LSystem Pattern Generation, Refactored Performance class into Timer class, Updated Bolt generation and handling
Added LSystem pattern geneartion. Updated the program to handle the extra choice and made GUI controls for it's behaviour: maxDisplacement and detial

Updated LightManager to have a max number of point lights and Updated lightPerSeg to have different values for the different methods. The LSystem gains an exponential number of segments and detail is increased so it's lightsPerSeg scales inversly with detail. (needs a look at for a proper value).

Updated BoltSetup to keep track of the number of segemnts in the static bolt patterns as well as the number of active lights. This was required because many functions were going throug the while array when they didn't need to.