Added GUI options, Updated scene, Updated geometry pass, Removed backpack model
Added debugging GUI window that lets you toggle the geometry shader to use diffuse, normal and specular textures and option to set the default diffuse color. Added option to set the number of segments for particle and random methods in the bolt generation window, just displays the number of segmetns for LSystem.

Added SetLSystemOptions to lightningPatterns which is called in main and sets the initial options for the LSystem. Updated geometry pass to only generate shadow maps after a new bolt is made instead of every frame. Fixed bug in ProjectBasePath which didn't work when running the release version.

Removed backpack model. Updated scene to have backwall.